Change DI to configure using IMvcCoreBuilder or IMvcBuilder #30
